chore: mv page files
