--- news_all/pages/news_all/index.vue
@@ -1,20 +1,0 @@
-<template>
-  <div>
-    <p>ニュース一覧ページ</p>
-    <div v-for="n in response.list" :key="n.slug">
-      <nuxt-link :to="`${$route.path}/${n.topics_id}`">
-        {{ n.ymd }} {{ n.subject }}
-      </nuxt-link>
-    </div>
-  </div>
-</template>
-
-<script>
-export default {
-  async asyncData({ $axios }) {
-    return {
-      response: await $axios.$get('/rcms-api/4/news'),
-    };
-  },
-};
-</script>
